feat: use cross-fetch as ponyfill
This means that users no longer need to think about polyfilling fetch in
their environment. This is especially helpful for node users.

This could be considered a breaking change because by taking a ponyfill
approach we're no longer relying on the host platform fetch
implementation, and that might lead to some unexpected behaviours. In
practice it should be fine given the battle tested libs under use here
and the simple use-case graphql-request has for them.